+++ .gitignore
@@ -1,3 +1,4 @@
 *.spec
 build/
 dist/
+__pycache__/
