+++ readme.md
@@ -1,0 +1,110 @@
+
+<!---------------------------------[ Header ]---------------------------------->
+
+<h1 align = 'center'>
+    <img 
+        src = 'assets/logo.png' 
+        height = '100' 
+        width = '100' 
+        alt = 'Icon' 
+    />
+    <br>
+    Findely - Technology
+    <br>
+</h1>
+
+<!---------------------------------[ Badges ]---------------------------------->
+
+<div align = 'center'>
+    <a href = 'https://github.com/Rohith0009/Findely/'>
+        <img src = 'https://img.shields.io/github/release-date/rohith0009/findely'/>
+    </a>
+    <a href = 'https://github.com/Rohith0009/Findely/'>
+        <img src = 'https://img.shields.io/github/downloads/rohith0009/findely/total'/>
+    </a>
+    <a href = 'https://github.com/Rohith0009/Findely/'>
+        <img src = 'https://img.shields.io/github/v/release/rohith0009/findely?include_prereleases'/>
+    </a>
+    <a href = 'https://github.com/Rohith0009/Findely/'>
+        <img src = 'https://img.shields.io/github/languages/top/rohith0009/findely'/>
+    </a>
+    <a href = 'https://github.com/Rohith0009/Findely/'>
+        <img src = 'https://img.shields.io/github/checks-status/rohith0009/findely/master'/>
+    </a>
+        <a href = 'https://github.com/Rohith0009/Findely/'>
+        <img src = 'https://img.shields.io/github/commits-since/rohith0009/findely/1'/>
+    </a>
+    
+</div>
+
+<br>
+
+### **What Is This?** - It Allows you to search and check if a word is there in a sentence or not. It also shows has a option to show how many times it is repeated. (see features below).
+
+### **How to Download:** Click the "[Releases]" link on the right, then on the latest release, under the latest version click to download "Findely.exe". (The name might have the release number in it.)
+
+## **Features**
+
+* ### 2 Different Modes:
+  * **Simple**: Does the Basics Of the App! It Just Scans And tells If The word is found or not.
+  * **Complex**: Does Way More The Than the Complex Mode! It can check how many times the word is repeated, Where it is located and more!
+
+
+## **Purpose**
+
+The main purpose of this app is to help you check if a word is present or not in a very big sentence and some more things!
+
+## **Usage Notes**
+
+1. **IF IT FREEZES** while running, it is probably because you clicked within the command prompt window which pauses everything. **To unfreeze it, simply-right click within the window, or press the Escape key.**
+
+2. I'm a total amateur, so if something doesn't work I'll try to fix it but might not even know how, so don't expect too much.
+
+## **Screenshots**
+
+<div align = 'center'>
+    <h2>Opening Menu</h2>
+    <br>
+    <img 
+        width = '675' 
+        alt = 'Opening Menu' 
+        src = 'https://i.ibb.co/17jpXBc/Opening-Menu.png"'>
+    <br>
+    <h2>Simple Mode</h2>
+    <br>
+    <img 
+        width = '675' 
+        alt = 'Simple Mode' 
+        src = 'https://i.ibb.co/GxGz3Mn/Simple-Mode.png'>
+    <br>
+    <h2>Scanning</h2>
+    <br>
+    <img 
+        width = '675' 
+        alt = 'Complex Mode' 
+        src = 'https://i.ibb.co/X4GM04w/complex.png'>
+    <br>
+</div>
+
+<br>
+
+## **Operating System Specific Instructions:**
+
+### Windows
+
+On Windows, you may download the latest EXE file from the [Releases page][Releases], or you may install Findely manually.
+If you are not sure how to install Findely manually, it is probably best to install the EXE.
+
+### Linux / MacOS
+As Of Now We Don't Support Linux Or MacOS
+
+## **Updating**
+
+To update Findely you need to download the latest EXE from the [Releases page][Releases] and replace it with The Old EXE
+
+<!----------------------------------------------------------------------------->
+
+[Releases]: https://github.com/rohith0009/findely/releases
+
+[Wiki]: https://github.com/rohith009/findely/wiki
+
